Adding async logging function
