+++ Dockerfile-msc
@@ -5,7 +5,7 @@ WORKDIR /usr/app
 COPY ms/Configuration-tmplt.java ./
 RUN sed "s/<PASSWORD>//g" < Configuration-tmplt.java > Configuration.java
 RUN rm Configuration-tmplt.java
-COPY ms/registry.properties ms/OrdersUI.java ms/MSClientAPI.java ms/CreateServicesAI.java ms/RetrieveServicesAI.java ms/Logger.java ./
+COPY ms/registry.properties ms/OrdersUI.java ms/MSClientAPI.java ms/CreateServicesAI.java ms/RetrieveServicesAI.java ms/ClientLogger.java ./
 RUN javac *.java
 
 CMD java OrdersUI
